[SPARK-37164][SQL] Add ExpressionBuilder for functions with complex overloads
### What changes were proposed in this pull request?

This PR adds a new trait `ExpressionBuilder` to help register functions with complex overloads that are hard to implement in the class constructor. And use this new API to register `lpad` and `rpad` functions, to fix the breaking change and allow these two functions to have mixed string and binary inputs under non-ANSI mode.

### Why are the changes needed?

In https://github.com/apache/spark/pull/34407 , we realized that the lpad and rpad functions are convoluted because:
1. We want to return binary if the `str` and `pad` are both binary.
2. We want to return string if the `str` and `pad` are both string.
3. If one parameter is string and the other is binary, we want to return string to not break existing queries. But with ansi mode, we want to fail as this case is likely a user mistake.

It's very hard to implement this logic with constructor, and #34407 chooses to let the function support mixed input types at runtime instead of compile time, which is a bit hacky.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

existing tests

Closes #34441 from cloud-fan/expression.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>