[SPARK-37191][SQL] Allow merging DecimalTypes with different precision values
### What changes were proposed in this pull request?

This PR updates the code in `StructType` to allow merging `DecimalType` values with different precision but the same scale.

We could allow merging DecimalType values with different precision if the scale is the same for both types since there should not be any data correctness issues as one of the types will be extended, for example, `DECIMAL(12, 2)` -> `DECIMAL(17, 2)`; however, this is not the case for upcasting when the scale is different - this would depend on the actual values.

This also affects Parquet schema merge which is where this issue was discovered originally.

### Why are the changes needed?

Fixes the issue of not allowing to consolidate different DecimalTypes when it should be fine to do so in case of the same scale value.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?

Added unit tests in StructTypeSuite and ParquetSchemaSuite.

Closes #34462 from sadikovi/SPARK-37191.

Authored-by: Ivan Sadikov <ivan.sadikov@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>