[SPARK-37702][SQL] Use AnalysisContext to track referred temp functions
### What changes were proposed in this pull request?
This PR uses `AnalysisContext` to track the referred temp functions in order to fix a temp
function resolution issue when it's registered with a `FunctionBuilder` and referred by a temp view.

During temporary view creation, we need to collect all the temp functions and save them
to the metadata. So that next time when resolving the view SQL text, the functions can be
resolved correctly. But if the temp function is registered with a `FunctionBuilder`, it's not a
`UserDefinedExpression` so it cannot be collected as a temp function. As a result, the next time
when the analyzer resolves a temp view, the registered function couldn't be found.

Example:
```scala
val func = CatalogFunction(FunctionIdentifier("tempFunc", None), ...)
val builder = (e: Seq[Expression]) => e.head
spark.sessionState.catalog.registerFunction(func, Some(builder))
sql("create temp view tv as select tempFunc(a, b) from values (1, 2) t(a, b)")
sql("select * from tv").collect()
```

### Why are the changes needed?
bug-fix

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
newly added test cases.

Closes #34546 from linhongliu-db/SPARK-37702-ver3.

Authored-by: Linhong Liu <linhong.liu@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>