[SPARK-37335][ML] Flesh out FPGrowth docs
### What changes were proposed in this pull request?

Some of the columns included in the DataFrames returned by `FPGrowthModel` are not explained. This PR adds some basic documentation for them. Since those docs depend on the concept of an _itemset_, this PR also introduces that term.

### Why are the changes needed?

The FPGrowth docs should cover at a high level all the basic information that the algorithm provides.

### Does this PR introduce _any_ user-facing change?

Yes, it updates user-facing documentation.

### How was this patch tested?

I will confirm the doc changes are correct as part of PR review.

Closes #34605 from nchammas/SPARK-37335-fpgrowth-docs.

Authored-by: Nicholas Chammas <nicholas.chammas@gmail.com>
Signed-off-by: Sean Owen <srowen@gmail.com>