[SPARK-36646][SQL] Push down group by partition column for aggregate
### What changes were proposed in this pull request?
lift the restriction for aggregate push down for parquet and orc if group by columns are the same as the partition cols

### Why are the changes needed?
previously, if there are group by columns, we don't push down aggregate to data source.
After the change, if the group by columns are the same as the partition columns, we will push down aggregates.

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
new tests

Closes #34445 from huaxingao/group_by.

Authored-by: Huaxin Gao <huaxin_gao@apple.com>
Signed-off-by: Liang-Chi Hsieh <viirya@gmail.com>