[SPARK-36346][SQL] Support TimestampNTZ type in Orc file source
### What changes were proposed in this pull request?
As per https://orc.apache.org/docs/types.html, ORC includes two different forms of timestamps from the SQL world:

`Timestamp` is a date and time without a time zone, which does not change based on the time zone of the reader.
`Timestamp with local time zone` is a fixed instant in time, which does change based on the time zone of the reader.

So the contrast relationship of timestamp between Spark and ORC as follows:
- `TIMESTAMP_LTZ` => `Timestamp with local time zone`.
- `TIMESTAMP_NTZ` => `Timestamp`

Unfortunately, in Spark 3.1 or prior, Spark considered ORC `Timestamp` as `TIMESTAMP_LTZ` mistakely.
The behavior is not correct. To keep backward compatibility, we not change the mistake now.
Since 3.2, with the support of timestamp without time zone type:

- For the ORC writer, we will have to map both Spark’s `TimestampLTZ` and `TimestampNTZ` as ORC’s `TimestampNTZ` for backward compatibility. In addition, the ORC should write the column type in ORC’s type information with the key “spark.sql.catalyst.type”.
- For the ORC reader. If the type information contains the key “spark.sql.catalyst.type”, infer the data type as per the value
Otherwise, given an ORC `TimestampLTZ` column, we will infer it as Spark’s `TimestampType`; given an ORC `TimestampNTZ`, the inferred result is `TimestampType`(`TimestampLTZ`) so that the behavior is consistent with older version Spark.

### Why are the changes needed?
Docking `TimestampNTZ` type in Spark and `Timestamp` in ORC.

### Does this PR introduce _any_ user-facing change?
'Yes'.
Spark will read/write the `TimestampNTZ`/`Timestamp` in ORC.

### How was this patch tested?
New tests.

Closes #33588 from beliefer/SPARK-36346.

Lead-authored-by: Jiaan Geng <beliefer@163.com>
Co-authored-by: gengjiaan <gengjiaan@360.cn>
Signed-off-by: Gengliang Wang <gengliang@apache.org>