[SPARK-37368][SQL][TESTS] Explicit GC for TPC-DS query runs
### What changes were proposed in this pull request?

This PR proposes to explicitly GC between each query of TPC-DS.

### Why are the changes needed?

Due to the lack of memory in GitHub Actions machines, they easily die in the middle of testing. This is rather a bandaid workaround.

### Does this PR introduce _any_ user-facing change?

No, dev-only.

### How was this patch tested?

It has been tested in the fork of Spark in the company I work for.

Closes #34641 from HyukjinKwon/SPARK-37368.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>