[SPARK-37224][SS][FOLLOWUP] Clarify the guide doc and fix the method doc
### What changes were proposed in this pull request?

This PR is a follow-up of #34502 to address post-reviews.

This PR rewords on the explanation on performance tune on RocksDB state store to make it less confused, and also fix the method docs to be in sync with the code changes.

### Why are the changes needed?

1. The explanation on performance tune on RocksDB state store was unclear in a couple of spots.
2. We changed the method signature, but the change was not reflected to the method doc.

### Does this PR introduce _any_ user-facing change?

Yes, end users will get less confused from the explanation on performance tune on RocksDB state store.

### How was this patch tested?

N/A

Closes #34652 from HeartSaVioR/SPARK-37224-follow-up-postreview.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>