[SPARK-37224][SS][FOLLOWUP] Add benchmark on basic state store operations
### What changes were proposed in this pull request?

This PR proposes to add a new benchmark to measure the performance on basic state store operations, and the result file.

The proposed change of SPARK-37224 (#34502) is applied in the benchmark. As the benchmark number provides, turning off the config brings lots of performance gain in micro-benchmark perspective, while it is still slower than memory-based state store.

### Why are the changes needed?

To track and verify further performance improvements.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Result file from manual run is included in this PR.

Closes #34630 from HeartSaVioR/SPARK-37224-follow-up-benchmark.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Liang-Chi Hsieh <viirya@gmail.com>