[SPARK-37388][SQL] Fix NPE in WidthBucket in WholeStageCodegenExec
### What changes were proposed in this pull request?

This PR fixes a `NullPointerException` in `WholeStageCodegenExec` caused by the expression `WidthBucket`. The cause of this NPE is that `WidthBucket` calls `WidthBucket.computeBucketNumber`, which can return `null`, but the generated code cannot deal with `null`s.

### Why are the changes needed?

This fixes a `NullPointerException` in Spark SQL.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Added tests to `MathExpressionsSuite`. This suite already had tests for `WidthBucket` with interval inputs, but lacked tests with double inputs. I checked that the tests failed without the fix, and succeed with the fix.

Closes #34670 from tomvanbussel/SPARK-37388.

Authored-by: Tom van Bussel <tom.vanbussel@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>