[SPARK-37454][SQL] Support expressions in time travel timestamp
### What changes were proposed in this pull request?

Allow expressions to be the timestamp of time travel spec, with 3 limitations:
1. the expression must be foldable (can't refer to columns)
2. can not contain subqueries
3. can be casted to timestamp

### Why are the changes needed?

It's convenient to support expressions as timestamp in time travel, e.g. we can query the table in yesterday by `FROM t TIMESTAMP AS OF current_date() - INTERVAL 1 DAY`, or `FROM t TIMESTAMP AS OF TIMESTAMP'yesterday'`.

### Does this PR introduce _any_ user-facing change?

yes, a new feature

### How was this patch tested?

new tests

Closes #34699 from cloud-fan/delta.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>