[SPARK-37311][SQL] Migrate ALTER NAMESPACE ... SET LOCATION to use V2 command by default
### What changes were proposed in this pull request?

This PR proposes to use V2 commands as default as outlined in [SPARK-36588](https://issues.apache.org/jira/browse/SPARK-36588), and this PR migrates `ALTER NAMESPACE ... SET LOCATION` to use v2 command by default.

Note that the work to make v1/v2 commands consistent was done in #34686 and tests covering both v1/v2 were done in #34610.

### Why are the changes needed?

It's been a while since we introduced the v2 commands,  and it seems reasonable to use v2 commands by default even for the session catalog, with a legacy config to fall back to the v1 commands.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing *AlterNamespaceSetLocationSuite tests cover this PR's change.

Closes #34708 from imback82/v2_alter_ns_location.

Authored-by: Terry Kim <yuminkim@gmail.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>