[SPARK-37389][SQL][FOLLOWUP] SET command shuold not parse comments
### What changes were proposed in this pull request?

This PR is a followup of https://github.com/apache/spark/pull/34668 , to fix a breaking change. The SET command uses wildcard which may contain unclosed comment, e.g. `/path/to/*`, and we shouldn't fail it.

This PR fixes it by skipping the unclosed comment check if we are parsing SET command.

### Why are the changes needed?

fix a breaking change

### Does this PR introduce _any_ user-facing change?

no, the breaking change is not released yet.

### How was this patch tested?

new tests

Closes #34763 from cloud-fan/set.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>