[SPARK-37529][K8S][TESTS][FOLLOWUP] Allow dev-run-integration-tests.sh to take a custom Dockerfile
### What changes were proposed in this pull request?

This PR changes `dev-run-integration-tests.sh` to allow it to take a custom Dockerfile like #34790 did.
With this change, this script accepts `--docker-file` option, which takes a path to a custom Dockerfile.
```
$ ./dev/run-integration-tests.sh --docker-file /path/to/dockerfile
```

### Why are the changes needed?

As of #34790, we can specify a custom Dockerfile by `spark.kubernetes.test.dockerFile` property  when we run the K8s integration tests using Maven.
We can run the integration test via `dev-run-integration-tests.sh` but there is no way to specify a custom Dockerfile.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Confirmed that the K8s integration tests run with the following command using `Dockerfile.java17`.
```
cd resource-managers/kubernetes/integration-tests
./dev/dev-run-integration-tests.sh --docker-file ../docker/src/main/dockerfiles/spark/Dockerfile.java17
```

Closes #34818 from sarutak/kube-integration-test-java17-2.

Authored-by: Kousuke Saruta <sarutak@oss.nttdata.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>