[SPARK-37561][SQL] Avoid loading all functions when obtaining hive's DelegationToken
### What changes were proposed in this pull request?
Use `Hive.getWithoutRegisterFns` instead of `Hive.get` to avoid loading all permanent functions from the hive meta store.

### Why are the changes needed?
At present, when obtaining the delegationToken of hive, all functions will be loaded.
This is unnecessary, it takes time to load the function, and it also increases the burden on the hive meta store.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
manual test

Closes #34822 from cxzl25/SPARK-37561.

Authored-by: sychen <sychen@ctrip.com>
Signed-off-by: Chao Sun <sunchao@apple.com>