[SPARK-32307][FOLLOW-UP][SQL] Set outputEncoder to None for ScalaUDF.preCanonicalized
### What changes were proposed in this pull request?

This PR proposes to set  `outputEncoder` to `None` for `ScalaUDF.preCanonicalized`.

### Why are the changes needed?

We once did the same thing to `inputEncoders` in https://github.com/apache/spark/pull/29106 to fix a bug where the canonicalized ScalaUDFs for the same ScalaUDF becomes different after resolving `inputEncoders`.  So this PR applies the same fix to `outputEncoder` to avoid hitting the same issue in the future. Note that we don't have the issue caused by `outputEncoder` now since we don't resolve `outputEncoder` yet.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass existing tests.

Closes #34937 from Ngone51/SPARK-32307-followup.

Authored-by: yi.wu <yi.wu@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>