[SPARK-36361][SPARK-37761][INFRA] Update GitHub Action image to 20211228
### What changes were proposed in this pull request?

This PR aims to update `GitHub Action` image to 20211228.

### Why are the changes needed?

This has `coverage` and `matplotlib` library. The following is the full diff from the previous images. Please note that I didn't rebuilt the Docker image from the scratch because it fails with some reasons. This image is built based on the previous 20211116 and installed `coverage` and `matplotlib` with their transitive dependencies.

**Python3.9**
- coverage==6.2
- cycler==0.11.0
- fonttools==4.28.5
- kiwisolver==1.3.2
- matplotlib==3.5.1
- Pillow==8.4.0

**PyPy3**
- coverage==6.2
- cycler==0.11.0
- fonttools==4.28.5
- kiwisolver==1.3.2
- matplotlib==3.5.1
- packaging==21.3
- Pillow==8.4.0
- pyparsing==3.0.6

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass the GitHub Action Jobs.

Closes #35044 from dongjoon-hyun/SPARK-36361_SPARK-37761.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>