[SPARK-37777][SQL] Update the SQL syntax of SHOW FUNCTIONS
### What changes were proposed in this pull request?

The SQL syntax of `SHOW FUNCTIONS` is very weird today. If you want to specify the database to list functions, you need to write `SHOW FUNCTIONS LIKE db.f1`. This is inconsistent with `SHOW TABLES` and `SHOW VIEWS`.

This PR proposes to follow `SHOW TABLES` and `SHOW VIEWS`, and support `SHOW FUNCTIONS FROM/IN db LIKE pattern`. To keep backward compatibility, the legacy syntax is still supported, but it can't be used together with the new `FROM/IN db` clause.

### Why are the changes needed?

Be consistent with `SHOW TABLES` and `SHOW VIEWS`

### Does this PR introduce _any_ user-facing change?

Yea, as it extends the `SHOW FUNCTIONS` SQL syntax.

### How was this patch tested?

updated tests

Closes #35056 from cloud-fan/show-functions.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>