[SPARK-37735][K8S] Add appId interface to KubernetesConf
### What changes were proposed in this pull request?
Add `appId` interface to KubernetesConf

### Why are the changes needed?
The `appId` now can be only accessed in `KuberntesDriverConf` and `KubernetesExecutorConf`, but can't be accesssed in `KubernetesConf`.

Some user featurestep are using `KubernetesConf` as init constructor parameter in order to share the featurestep between driver and executor. One of cases is customized feature step (such as volcano, yunikorn) is using appId as job identification.

So, we'd better add appId to KubernetesConf to help such featurestep access appId.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
UT

Closes #35015 from Yikun/SPARK-36057.

Authored-by: Yikun Jiang <yikunkero@gmail.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>