[SPARK-37817][SQL][MINOR] Remove unreachable code in complexTypeExtractors.scala
### What changes were proposed in this pull request?

The constructor of GetArrayItem and GetMapValue with two input elements are not actually used. Usually, we implement such construction method so that it can be used in FunctionRegistry
```
def this(child: Expression, ordinal: Expression) = this(child, ordinal, SQLConf.get.ansiEnabled)
```

Since the expression GetArrayItem and GetMapValue are not used as functions in FunctionRegistry, we can remove these constructors.

### Why are the changes needed?

Code clean up

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing UT

Closes #35106 from gengliangwang/removeUnusedCode.

Authored-by: Gengliang Wang <gengliang@apache.org>
Signed-off-by: Gengliang Wang <gengliang@apache.org>