[SPARK-36879][SQL] Support Parquet v2 data page encoding (DELTA_BINARY_PACKED) for the vectorized path
### What changes were proposed in this pull request?
Implements a vectorized version of the parquet reader for DELTA_BINARY_PACKED encoding
This PR includes a previous PR for this issue which passed the read request thru to the parquet implementation and which was not vectorized. The current PR builds on top of that PR (hence both are included).

### Why are the changes needed?
Currently Spark throws an exception when reading data with these encodings if vectorized reader is enabled

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Additional unit tests for the encoding for both long and integer types (mirroring the unit tests in the Parquet implementation)

Closes #34471 from parthchandra/SPARK-36879-PR1.

Lead-authored-by: Parth Chandra <parthc@apache.org>
Co-authored-by: Nandor Kollar <nkollar@cloudera.com>
Signed-off-by: Chao Sun <sunchao@apple.com>